refactor(api/p1): split app.ts into context + lib + per-feature route modules
app.ts went from 2,263 lines (75 routes, one file) to ~120: it now builds a
shared AppContext, applies middleware (A4 sandbox-origin, CORS, rate limit, auth),
and mounts one Hono router per feature.

- context.ts: buildContext(deps) carries the shared closures (auth, authz,
  workspace resolution, quotas/limits, realtime bus, notify, sourceText).
  AppContext = ReturnType<typeof buildContext>, so the type can't drift.
- lib/: pure helpers — crypto (sha256/safeEqual), rate-limit, net (SSRF guard),
  http (headers/constants/cors/validators), comments (meta/mentions/format).
- routes/: session, workspace, agents, artifacts, sharing, favorites, collections,
  moderation, proposals, comments, realtime, analytics, notifications, webhooks,
  viewer, raw — each a (ctx) => Hono factory.

Pure restructure: handler bodies moved verbatim. node.ts unchanged. Public surface
(createApp, AppDeps) preserved. Gate green: typecheck, biome 0 errors, 131 api tests.
