+++ apps/api/src/lib/crypto.ts
@@ -1,0 +1,10 @@
+import { createHash, timingSafeEqual } from "node:crypto"
+
+export const sha256 = (s: string): string => createHash("sha256").update(s).digest("hex")
+
+// Constant-time string compare: hash both sides to a fixed length first so
+// neither the contents nor the length leak through comparison timing. An unset
+// secret (undefined) never matches.
+export const safeEqual = (a: string, b: string | undefined): boolean =>
+  !!b &&
+  timingSafeEqual(createHash("sha256").update(a).digest(), createHash("sha256").update(b).digest())
