+++ apps/api/src/lib/http.ts
@@ -1,0 +1,89 @@
+import type { Role, Visibility } from "@dock/core"
+import type { Context } from "hono"
+
+export const DEFAULT_WORKSPACE_NAME = "My Workspace"
+/** Cookie holding the active workspace id (multi-workspace mode). */
+export const WS_COOKIE = "dock_ws"
+
+export const VISIBILITIES = ["public", "link", "org"] as const
+
+export const MAX_UPLOAD_BYTES = 100 * 1024 * 1024
+
+/** Repeat opens by the same viewer of the same version inside this window collapse
+ *  to one recorded view — a refresh or quick re-open doesn't inflate the count. */
+export const VIEW_DEDUP_MS = 30 * 60_000
+
+/** Headers for everything inside the artifact sandbox. */
+export const RAW_HEADERS: Record<string, string> = {
+  // Opaque origin: scripts run, but can touch no cookies, storage, or APIs.
+  "Content-Security-Policy":
+    "sandbox allow-scripts allow-forms allow-popups allow-modals allow-downloads",
+  "Access-Control-Allow-Origin": "*",
+  "X-Content-Type-Options": "nosniff",
+  // Raw artifact bytes are never the indexable surface (the viewer is); keeping
+  // them out of search engines also blunts using the host for SEO-spam/phishing.
+  "X-Robots-Tag": "noindex",
+  // Versioned paths are immutable by construction.
+  "Cache-Control": "public, max-age=31536000, immutable",
+}
+
+export const REF_RE = /^([0-9a-z]{6,12})(?:-[a-z0-9-]*)?(?:@v(\d+))?$/
+
+/** A taken-down artifact: content is gone (410), the record is preserved. */
+export const TOMBSTONE = "This artifact was removed."
+
+/** Copy into a plain ArrayBuffer — what Hono's body() accepts. */
+export const toBody = (u: Uint8Array): ArrayBuffer => new Uint8Array(u).buffer as ArrayBuffer
+
+/**
+ * Embedded mode serves bundles under /raw/:id/v/:n/, so root-absolute URLs
+ * (href="/x", src="/x", url(/x)) must be prefixed or they escape the artifact.
+ * Domain mode (per-artifact origin) makes this a no-op later.
+ */
+export const rewriteAbsoluteUrls = (text: string, prefix: string): string =>
+  text
+    .replace(/(\b(?:href|src|action|srcset|poster)=["'])\/(?!\/)/g, `$1${prefix}/`)
+    .replace(/(url\(\s*['"]?)\/(?!\/)/g, `$1${prefix}/`)
+
+// Workspace membership is three simple roles, presented to people as
+// Admin / Creator / Viewer:
+//   - owner     → Admin:   manage members + settings (and everything below)
+//   - editor    → Creator: create + publish artifacts
+//   - commenter → Viewer:  read + comment
+// The canonical Role vocabulary is unchanged; a bare read-only "viewer" isn't
+// offered as a workspace role (a Viewer can always comment).
+export const isWorkspaceRole = (v: unknown): v is Role =>
+  v === "owner" || v === "editor" || v === "commenter"
+
+export const str = (v: unknown): string | undefined =>
+  typeof v === "string" && v !== "" ? v : undefined
+
+export const visibilityOf = (v: unknown): Visibility | undefined =>
+  typeof v === "string" && (VISIBILITIES as readonly string[]).includes(v)
+    ? (v as Visibility)
+    : undefined
+
+/**
+ * Echo-origin CORS middleware so the cross-origin SPA can send cookies. Headers
+ * are written onto the final response after next() — the Better Auth handler
+ * returns its own Response, so setting them beforehand would be discarded.
+ */
+export const corsFor = (allowed: Set<string>) => async (c: Context, next: () => Promise<void>) => {
+  const origin = c.req.header("origin")
+  const ok = !!origin && allowed.has(origin)
+  if (ok && c.req.method === "OPTIONS")
+    return c.body(null, 204, {
+      "Access-Control-Allow-Origin": origin,
+      "Access-Control-Allow-Credentials": "true",
+      "Access-Control-Allow-Methods": "GET,POST,OPTIONS",
+      "Access-Control-Allow-Headers": "content-type,authorization",
+      "Access-Control-Max-Age": "86400",
+      Vary: "Origin",
+    })
+  await next()
+  if (ok) {
+    c.res.headers.set("Access-Control-Allow-Origin", origin)
+    c.res.headers.set("Access-Control-Allow-Credentials", "true")
+    c.res.headers.append("Vary", "Origin")
+  }
+}
