+++ apps/api/src/routes/notifications.ts
@@ -1,0 +1,55 @@
+import { Hono } from "hono"
+import { streamSSE } from "hono/streaming"
+import type { AppContext } from "../context"
+
+/** In-app notifications (the header bell) for the signed-in user. */
+export const notificationRoutes = (ctx: AppContext) => {
+  const { meta, bus, currentUser } = ctx
+  const app = new Hono()
+
+  app.get("/v1/notifications", async (c) => {
+    const me = await currentUser(c)
+    if (!me) return c.json({ error: "unauthenticated" }, 401)
+    const [notifications, unread] = await Promise.all([
+      meta.listNotifications(me.id, 50),
+      meta.unreadNotificationCount(me.id),
+    ])
+    return c.json({ notifications, unread })
+  })
+
+  app.post("/v1/notifications/read", async (c) => {
+    const me = await currentUser(c)
+    if (!me) return c.json({ error: "unauthenticated" }, 401)
+    const body = (await c.req.json().catch(() => ({}))) as { ids?: unknown; all?: unknown }
+    const ids =
+      body.all === true
+        ? "all"
+        : Array.isArray(body.ids)
+          ? body.ids.filter((x): x is string => typeof x === "string")
+          : []
+    await meta.markNotificationsRead(me.id, ids)
+    const unread = await meta.unreadNotificationCount(me.id)
+    return c.json({ unread })
+  })
+
+  // Live notification stream for the signed-in user (the header bell subscribes).
+  app.get("/v1/notifications/events", async (c) => {
+    const me = await currentUser(c)
+    if (!me) return c.text("unauthenticated", 401)
+    c.header("Access-Control-Allow-Origin", "*")
+    const userId = me.id
+    return streamSSE(c, async (stream) => {
+      const unsub = bus.subscribe(`u:${userId}`, (e) => {
+        void stream.writeSSE({ event: e.type, data: JSON.stringify(e) })
+      })
+      stream.onAbort(unsub)
+      await stream.writeSSE({ event: "ready", data: "{}" })
+      while (!stream.aborted) {
+        await stream.sleep(15000)
+        await stream.writeSSE({ event: "ping", data: "{}" })
+      }
+    })
+  })
+
+  return app
+}
